[Update] 2473 Two Pointer

+++ src/BaekJoon/2473.cpp
@@ -9,27 +9,27 @@ void Init()
 	std::cin.tie(NULL); std::cout.tie(NULL);
 }
 
-std::vector<int> data;
-int N, from = 0, to = 1, L = 0, M = 0, R = 0;
+std::vector<long long> data;
+long long N, from = 0, to = 1, L = 0, M = 0, R = 0;
 long long min = 3e9 + 1;
 
-int func(int from, int to, int current)
+long long func(long long from, long long to, long long current)
 {
 	if (from == to) return from;
 
 	if (from + 1 == to)
 	{
-		int F = std::abs(data[from] + current);
-		int T = std::abs(data[to] + current);
+		long long F = std::abs(data[from] + current);
+		long long T = std::abs(data[to] + current);
 
 		if (F > T) return to;
 		else return from;
 	}
 
-	int mid = (from + to) / 2;
+	long long mid = (from + to) / 2;
 
-	int L = func(from, mid, current);
-	int R = func(mid + 1, to, current);
+	long long L = func(from, mid, current);
+	long long R = func(mid + 1, to, current);
 	long long left = std::abs(data[L] + current);
 	long long right = std::abs(data[R] + current);
 
@@ -42,8 +42,8 @@ int main()
 	std::cin >> N;
 
 	to = N - 1;
-	int tmp;
-	for (int i = 0; i < N; ++i)
+	long long tmp;
+	for (long long i = 0; i < N; ++i)
 	{
 		std::cin >> tmp;
 		data.push_back(tmp);
@@ -55,7 +55,7 @@ int main()
 	{
 		long long current = data[from] + data[to];
 
-		int mid = func(from + 1, to - 1, current);
+		long long mid = func(from + 1, to - 1, current);
 
 		current += data[mid];
 		if (std::abs(min) > std::abs(current))
